+++ src/libambulant/net/rtsp_factory.cpp
@@ -76,7 +76,7 @@ live_audio_datasource_factory::new_audio_datasource(const net::url& url, const a
 	AM_DBG lib::logger::get_logger()->debug("live_audio_datasource_factory::new_audio_datasource: decoder ds = 0x%x", (void*)dds);
 	if (dds == NULL) {
 		pds->stop();
-		int rem = pds->release();
+		long rem = pds->release();
 		assert(rem == 0);
 		return NULL;
 	}
@@ -88,7 +88,7 @@ live_audio_datasource_factory::new_audio_datasource(const net::url& url, const a
 	AM_DBG lib::logger::get_logger()->debug("live_audio_datasource_factory::new_audio_datasource: resample ds = 0x%x", (void*)rds);
 	if (rds == NULL)  {
 		dds->stop();
-		int rem = dds->release();
+		long rem = dds->release();
 		assert(rem == 0);
 		return NULL;
 	}
@@ -98,7 +98,7 @@ live_audio_datasource_factory::new_audio_datasource(const net::url& url, const a
 	}
 	lib::logger::get_logger()->error(gettext("%s: unable to create audio resampler"));
 	rds->stop();
-	int rem = rds->release();
+	long rem = rds->release();
 	assert(rem == 0);
 	return NULL;
 }
